get font family from theme insted of default fonts
